moved colorbar functions to utils
